+++ src/components/LineChart/components/Line/components/Point/Point.tsx
@@ -2,6 +2,8 @@ import React from 'react';
 import tokens from '@shopify/polaris-tokens';
 import {Color} from 'types';
 
+import {getColorValue} from '../../../../../../utilities';
+
 interface Props {
   active: boolean;
   cx: number;
@@ -15,7 +17,7 @@ export function Point({cx, cy, active, color}: Props) {
       cx={cx}
       cy={cy}
       r={active ? 5 : 0}
-      fill={tokens[color]}
+      fill={getColorValue(color)}
       stroke={tokens.colorWhite}
       strokeWidth={1.5}
     />